# apps/backend/core-api/docs/swagger.yaml
definitions:
...
  core-api_internal_handler_event.EventResponse:
    properties:
      attendees_count:
        type: integer
      banner_storage_key:
        type: string
      chain_id:
        type: integer
      contact_address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      event_status:
        $ref: '#/definitions/entity.EventStatus'
      event_type:
        $ref: '#/definitions/entity.EventType'
      google_map_query:
        type: string
      icon_storage_key:
        type: string
      id:
        type: string
      is_booking_request_required:
        type: boolean
      is_public:
        type: boolean
      is_ticket_transferable:
        type: boolean
      is_verified:
        type: boolean
      location:
        type: string
      long_description:
        type: string
      max_attendees:
        type: integer
      owner_credential_id:
        type: string
      short_description:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - attendees_count
    - banner_storage_key
    - chain_id
    - contact_address
    - contact_number
    - created_at
    - end_date
    - event_status
    - event_type
    - google_map_query
    - icon_storage_key
    - id
    - is_booking_request_required
    - is_public
    - is_ticket_transferable
    - is_verified
    - location
    - long_description
    - max_attendees
    - owner_credential_id
    - short_description
    - start_date
    - title
    - updated_at
    type: object
...
  customerror.ErrResponse:
    description: Response for the client to sign to register
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  entity.Event:
    properties:
      attendees_count:
        type: integer
      banner_storage_key:
        type: string
      chain_id:
        type: integer
      contact_address:
        type: string
      contact_number:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      event_status:
        $ref: '#/definitions/entity.EventStatus'
      event_type:
        $ref: '#/definitions/entity.EventType'
      google_map_query:
        type: string
      icon_storage_key:
        type: string
      id:
        type: string
      is_booking_request_required:
        type: boolean
      is_public:
        type: boolean
      is_ticket_transferable:
        type: boolean
      is_verified:
        type: boolean
      location:
        type: string
      long_description:
        type: string
      max_attendees:
        type: integer
      owner_credential_id:
        type: string
      short_description:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - attendees_count
    - banner_storage_key
    - chain_id
    - contact_address
    - contact_number
    - created_at
    - end_date
    - event_status
    - event_type
    - google_map_query
    - icon_storage_key
    - id
    - is_booking_request_required
    - is_public
    - is_ticket_transferable
    - is_verified
    - location
    - long_description
    - max_attendees
    - owner_credential_id
    - short_description
    - start_date
    - title
    - updated_at
    type: object
...
  entity.EventStatus:
    enum:
    - active
    - inactive
    - closed
    type: string
    x-enum-varnames:
    - EventStatusActive
    - EventStatusInactive
    - EventStatusClosed
  entity.EventType:
    enum:
    - private
    - invite
    type: string
    x-enum-varnames:
    - EventTypePrivate
    - EventTypeInvite
...
  entity.Profile:
    properties:
      academic_email:
        type: string
      academic_institution:
        type: string
      address:
        type: string
      authentication_credential_id:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      github_connector_ref:
        type: string
      google_connector_ref:
        description: Connector references (from authentication_credentials table)
        type: string
      id:
        type: string
      is_academic_email_public:
        type: boolean
      is_academic_institution_public:
        type: boolean
      is_address_public:
        type: boolean
      is_bio_public:
        type: boolean
      is_email_public:
        type: boolean
      is_first_name_public:
        type: boolean
      is_last_name_public:
        type: boolean
      is_phone_number_public:
        type: boolean
      is_profile_picture_public:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      profile_picture_url:
        type: string
      updated_at:
        type: string
      wallet_address:
        type: string
    required:
    - authentication_credential_id
    - created_at
    - id
    - is_academic_email_public
    - is_academic_institution_public
    - is_address_public
    - is_bio_public
    - is_email_public
    - is_first_name_public
    - is_last_name_public
    - is_phone_number_public
    - is_profile_picture_public
    - updated_at
    type: object
...
  event.CreateEventIssuerRequest:
    properties:
      is_signed:
        enum:
        - 0
        - 1
        type: integer
      issuer_credential_id:
        type: string
    required:
    - is_signed
    - issuer_credential_id
    type: object
...
  event.EventIssuerResponse:
    properties:
      created_at:
        type: string
      event_id:
        type: string
      id:
        type: string
      is_signed:
        type: integer
      issuer_credential_id:
        type: string
      issuer_profile:
        $ref: '#/definitions/entity.Profile'
      updated_at:
        type: string
    required:
    - created_at
    - event_id
    - id
    - is_signed
    - issuer_credential_id
    - updated_at
    type: object
  event.EventParticipantResponse:
    properties:
      academic_institution:
        type: string
      attendee_credential_id:
        type: string
      contract_address:
        type: string
      created_at:
        type: string
      email:
        type: string
      event_id:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_attendee_accepted:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      wallet_address:
        type: string
    required:
    - attendee_credential_id
    - contract_address
    - created_at
    - event_id
    - id
    - is_attendee_accepted
    - updated_at
    - wallet_address
    type: object
...
  event.GetCertificatesListViewModelResponse:
    properties:
      claimed_certificates: {}
      total_claimed:
        type: integer
      total_unclaimed:
        type: integer
      unclaimed_certificates: {}
    required:
    - claimed_certificates
    - total_claimed
    - total_unclaimed
    - unclaimed_certificates
    type: object
...
  event.GetEventListResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/entity.Event'
        type: array
    required:
    - events
    type: object
...
  event.UpdateEventIssuerRequest:
    properties:
      event_id:
        type: string
      issuer_credential_id:
        type: string
    required:
    - event_id
    - issuer_credential_id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: DECM (Decentralized Event Management) platform API for NFT ticketing,
    digital credentials, and academic identity verification.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DECM Core
  version: "1.0"
paths:
...
  /api/v1/events:
    get:
      description: Get events list
      operationId: get-events-list
      parameters:
      - default: true
        description: Include active events
        in: query
        name: include_active_events
        type: boolean
      - default: true
        description: Include inactive events
        in: query
        name: include_inactive_events
        type: boolean
      - default: false
        description: Include closed events
        in: query
        name: include_closed_events
        type: boolean
      - default: false
        description: Only user joined events
        in: query
        name: only_user_joined_events
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.GetEventListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get events list
      tags:
      - Event
    post:
      consumes:
      - multipart/form-data
      description: Create a new event with banner image upload. Chain ID is automatically
        set from environment configuration.
      operationId: create-event
      parameters:
      - description: Event name
        in: formData
        name: name
        required: true
        type: string
      - description: Event short description
        in: formData
        name: short_description
        required: true
        type: string
      - description: Event description
        in: formData
        name: description
        required: true
        type: string
      - description: Start date (RFC3339 format)
        in: formData
        name: start_date
        required: true
        type: string
      - description: End date (RFC3339 format)
        in: formData
        name: end_date
        required: true
        type: string
      - description: Number of seats
        in: formData
        name: seats_count
        required: true
        type: integer
      - description: Contact number
        in: formData
        name: contact_number
        required: true
        type: string
      - description: Contact address
        in: formData
        name: contact_address
        required: true
        type: string
      - description: Location
        in: formData
        name: location
        required: true
        type: string
      - description: Google map query
        in: formData
        name: google_map_query
        required: true
        type: string
      - description: Event banner image (JPEG, PNG, WebP, max 10MB)
        in: formData
        name: banner
        required: true
        type: file
      - description: Event icon image (JPEG, PNG, WebP, max 10MB)
        in: formData
        name: icon
        required: true
        type: file
      - description: Host password
        in: formData
        name: host_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/core-api_internal_handler_event.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Create a new event
      tags:
      - Event
...
  /api/v1/events/{event_id}/certificates/list-viewmodel:
    get:
      consumes:
      - application/json
      description: Get event certificates separated by claimed and unclaimed status.
        Claimed certificates have token_id populated, unclaimed certificates have
        token_id null and certificate config is published. Only event hosts and issuers
        can access this endpoint.
      operationId: get-certificates-list-viewmodel
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.GetCertificatesListViewModelResponse'
        "400":
          description: Invalid event ID
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized - missing or invalid authentication
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden - user is not host or issuer of this event
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get certificates list viewmodel
      tags:
      - Event Certificates
...
  /api/v1/events/{event_id}/issuers:
    get:
      consumes:
      - application/json
      description: Get all event issuers for an event
      operationId: get-event-issuers-by-event-id
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/event.EventIssuerResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get event issuers by event ID
      tags:
      - Events
    post:
      consumes:
      - application/json
      description: Create a new event issuer for an event
      operationId: create-event-issuer
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Event issuer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.CreateEventIssuerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventIssuerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Create event issuer
    put:
      consumes:
      - application/json
      description: Update an event issuer
      operationId: update-event-issuer
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Event issuer data
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/event.UpdateEventIssuerRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/event.EventIssuerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Update event issuer
...
  /api/v1/events/{event_id}/participants:
    get:
      consumes:
      - application/json
      description: Get all participants (attendees) for an event
      operationId: get-event-participants
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/event.EventParticipantResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.ErrResponse'
      summary: Get event participants
      tags:
      - Events
...
swagger: "2.0"